Remove largely incorrect usage of floating_to_fixed_point()
Added new functions float_to_color() and color_multiply() for dealing
with uint8_t colour values directly, with an assumed mapping from
0-255 to 0.0-1.0 and back.

+++ src/units/drawer.hpp
@@ -75,7 +75,7 @@ private:
 	/** draw a health/xp bar of a unit */
 	void draw_bar(const std::string& image, int xpos, int ypos,
 		const map_location& loc, std::size_t height, double filled,
-		const color_t& col, int32_t alpha) const;
+		const color_t& col, uint8_t alpha) const;
 
 	/**
 	 * Finds the start and end rows on the energy bar image.
